stats/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0,<4.0
mysqlclient>=1.4.0
pika
celery